Separating UI and other classes.
